Add meta row
