+++ components/brave_news/browser/resources/feed/Article.tsx
@@ -2,16 +2,14 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
-import * as React from 'react';
-import Card, { MetaInfo, Title, getOrigin } from './Card';
-import { FeedItemMetadata, Article as Info } from 'gen/brave/components/brave_news/common/brave_news.mojom.m';
-import styled from 'styled-components';
 import { spacing } from '@brave/leo/tokens/css';
-import { useLazyUnpaddedImageUrl } from '../../../../brave_new_tab_ui/components/default/braveNews/useUnpaddedImageUrl';
+import { FeedItemMetadata, Article as Info } from 'gen/brave/components/brave_news/common/brave_news.mojom.m';
+import * as React from 'react';
+import styled from 'styled-components';
 import Flex from '../../../../brave_new_tab_ui/components/Flex';
-import Button from '@brave/leo/react/button';
-import ButtonMenu from '@brave/leo/react/buttonMenu'
-import Icon from '@brave/leo/react/icon';
+import { useLazyUnpaddedImageUrl } from '../../../../brave_new_tab_ui/components/default/braveNews/useUnpaddedImageUrl';
+import ArticleMetaRow from './ArticleMetaRow';
+import Card, { Title } from './Card';
 
 interface Props {
   info: Info
@@ -33,35 +31,13 @@ const ArticleImage = styled.img`
 
   border-radius: 6px;
 `
-
-const MenuButton = styled(Button)`
-  flex-grow: 0;
-  --leo-button-padding: ${spacing.s};
-`
-
 export const openArticle = (article: FeedItemMetadata) => window.open(article.url.url, '_blank', 'noopener noreferrer')
 
 export default function Article({ info, hideChannel }: Props) {
   const { url, setElementRef } = useLazyUnpaddedImageUrl(info.data.image.paddedImageUrl?.url, { useCache: true })
 
   return <Container onClick={() => openArticle(info.data)} ref={setElementRef}>
-    <Flex direction='row' gap={spacing.m} justify='space-between'>
-      <MetaInfo article={info.data} hideChannel={hideChannel} />
-      <ButtonMenu>
-        <MenuButton slot='anchor-content' kind='plain-faint' onClick={e => {
-          e.preventDefault()
-          e.stopPropagation()
-          e.stopImmediatePropagation()
-        }}>
-          <Icon name='more-horizontal' />
-        </MenuButton>
-        <leo-menu-item onClick={e => {
-          e.preventDefault()
-          e.stopPropagation()
-          e.stopImmediatePropagation()
-        }}>Hide content from {getOrigin(info.data)}</leo-menu-item>
-      </ButtonMenu>
-    </Flex>
+    <ArticleMetaRow article={info.data} hideChannel={hideChannel} />
     <Flex direction='row' gap={spacing.m} justify='space-between'>
       <Title>{info.data.title}{('isDiscover' in info && info.isDiscover) && " (discovering)"}</Title>
       <ArticleImage src={url} />
