+++ components/brave_news/browser/resources/feed/Card.tsx
@@ -2,11 +2,9 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import { color, font, radius, spacing } from '@brave/leo/tokens/css';
 import styled from "styled-components";
-import { color, font, radius, spacing } from '@brave/leo/tokens/css'
-import { FeedItemMetadata } from "../../../../brave_new_tab_ui/api/brave_news";
-import { channelIcons } from "../../../../brave_new_tab_ui/components/default/braveNews/customize/Icons";
-import * as React from "react";
 
 export const Header = styled.h2`
   margin: 0;
@@ -16,36 +14,6 @@ export const Header = styled.h2`
 
   --leo-icon-size: 18px;
 `
-
-export const MetaInfoContainer = styled.h4` 
-  margin: 0;
-
-  font: ${font.primary.xSmall.regular};
-  color: ${color.text.secondary};
-
-  opacity: 0.5;
-
-  display: flex;
-  align-items: center;
-  gap: ${spacing.s};
-
-  --leo-icon-size: 12px;
-`
-
-export const getOrigin = (article: FeedItemMetadata) => {
-  const host = new URL(article.url.url).host
-  return host.startsWith('www.') ? host.substring(4) : host
-}
-
-export function MetaInfo(props: { article: FeedItemMetadata, hideChannel?: boolean }) {
-
-  const maybeChannel = !props.hideChannel && <>
-    • {channelIcons[props.article.categoryName] ?? channelIcons.default} {props.article.categoryName}
-  </>
-  return <MetaInfoContainer>
-    {getOrigin(props.article)} {maybeChannel} • {props.article.relativeTimeDescription}
-  </MetaInfoContainer>
-}
 
 export const Title = styled.h3`
   margin: 0;
