+++ components/brave_news/browser/resources/feed/Cluster.tsx
@@ -4,11 +4,12 @@
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 import * as React from 'react';
 import { Cluster as Info } from 'gen/brave/components/brave_news/common/brave_news.mojom.m';
-import Card, { MetaInfoContainer } from './Card';
+import Card from './Card';
 import Article from './Article';
 import styled from 'styled-components';
 import { spacing } from '@brave/leo/tokens/css';
 import { channelIcons } from '../../../../brave_new_tab_ui/components/default/braveNews/customize/Icons';
+import { MetaInfoContainer } from './ArticleMetaRow';
 
 interface Props {
   info: Info
